Fix checking for IOBase, and also set flash_addr. Fix handling of Descriptors
git-svn-id: https://svn.aros.org/svn/aros/trunk@30410 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/devs/networks/e1000/e1000_init.c
@@ -315,7 +315,7 @@ D(bug("[e1000] OpenDevice: Unit %d @ %p\n", unitnum, unit));
         else
         {
 D(bug("[e1000] OpenDevice: Starting Unit %d\n", unitnum));
-            unsigned int rx_ring_count, tx_ring_count;
+            ULONG rx_ring_count, tx_ring_count;
             e1000_mac_type mac_type;
 
             mac_type = ((struct e1000_hw *)unit->e1ku_Private00)->mac.type;
